requirement.txt
python==3.10
streamlit==1.26.0
pandas==2.1.1
matplotlib==3.7.2
arabic-reshaper==2.1.4
python-bidi==0.4.2